Full commit

+++ src/Tasks/Task_14102024/PrintMyBook.java
@@ -2,7 +2,7 @@ package Tasks.Task_14102024;
 
 public class PrintMyBook extends Book {
 
-    public PrintMyBook(String name, String author, int price) {
+    public PrintMyBook(String name, String author, double price) {
 
         super.name=name;
         super.author=author;
@@ -11,7 +11,7 @@ public class PrintMyBook extends Book {
 
     @Override
     void getdetails() {
-        System.out.printf("name = %s , author =%s, price = %d",name, author, price);
+        System.out.printf("name = %s , author =%s, price = %f",name, author, price);
 
     }
 }
